+++ src/main/java/com/jasso/finance/finance/service/BudgetService.java
@@ -5,9 +5,9 @@ import com.jasso.finance.finance.entity.Budget;
 import java.util.List;
 
 public interface BudgetService {
-    List<Budget> findAll(Integer id);
+    List<Budget> findAll(Long id);
     
-    List<Budget> findByUserIdAndMonthAndYear(Integer userId, Integer month, Integer year);
+    List<Budget> findByUserIdAndMonthAndYear(Long userId, Integer month, Integer year);
 
     Budget findById(Integer id);
