+++ src/main/java/com/jasso/finance/finance/service/TransactionService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface TransactionService {
     List<Transaction> findAll();
     
-    List<Transaction> findByUserId(Integer userId);
+    List<Transaction> findByUserId(Long userId);
 
     Transaction findById(Integer id);
